sonarcloud warnings 9 (#1511)
* refactor: const correctness 

* refactor: fix some implicit conversions

* refactor: make local pointers const if their objects are not modified

* refactor: do not cast away const in torrent-cell-renderer

* refactor: remove call to deprecated gtk_icon_size_lookup_for_settings

* refactor: member functions that do not mutate their objects should be declared const

* chore: do not end comments with a semicolon

+++ qt/FileTreeItem.cc
@@ -22,7 +22,7 @@
 
 QHash<QString, int> const& FileTreeItem::getMyChildRows()
 {
-    size_t const n = childCount();
+    int const n = childCount();
 
     // ensure that all the rows are hashed
     while (first_unhashed_row_ < n)
@@ -50,7 +50,7 @@ FileTreeItem::~FileTreeItem()
 
 void FileTreeItem::appendChild(FileTreeItem* child)
 {
-    size_t const n = childCount();
+    int const n = childCount();
     child->parent_ = this;
     children_.append(child);
     first_unhashed_row_ = n;
